Adding fwaas settings in neutron conf files (bsc#1113816)
Co-Authored-By: Charu Jain <cjain@suse.com>

Issues: SCRD-6071

Configure vpnaas as part of the l3 agent, and combine the setup of
the vpnaas and fwaas as part of the neutron-l3-agent role. This
allows us to remove the unused neutron-vpn-agent role, and any
references to config file templates associated with the older
implementation of vpnaas deployment.

Change-Id: Ie092d70027b310d378433bee39066118d5b5a1fc