SCRD-5854 Always configure fwaas and vpnaas l3-agent (SCRD-5854)
Always configure the fwaas and vpnaas extensions in any instance
of neutron-l3-agent that we start as empirical testing is showing
that the an l3-agent running on a contoller with only the vpnaas
extension enabled is being asked to handle fwaas operations. And
it appears the reverse is true for l3-agent instances that only
have the fwaas extension enabled, which appear to be being asked
to handle vpnaas operations.

Change-Id: I32be9a94c886bcf088e52d8dd154aa6d4f4f95ea