Improve neutron service restart limit handling (SOC-8746)
The default systemd restart limit handling will stop restarting a
service if it is restart more than 5 times in a 10 second interval.

However when we introduced the RestartSec setting, defaulting to
5 seconds, we also broke restart limiting, since with a 5 second
delay between restarts we could never restart more than twice in
a 10 second interval.

This patch adds support for configuring StartLimitInterval and
StartLimitBurst in addition to the RestartSec setting, defaulting
the interval to 40 seconds and the burst limit to 5.

Change-Id: I6b0a18055feaef38c624e433c362fce5ee4a0e4f