Add dependent services to neutron services (SOC-8746)
This patch addes "openvswitch.service" to the systemd service
definition of neutron-openvswitch-agent so that openvswitch (and its
dependent services) must be running before starting the agent.
Otherwise, the neutron-openvswitch-agent will continually crash
and restart if one or more sub-components of openvswitch is
stopped.

This change will also permit the configuration playbooks of other
neutron services to specify services that must be started prior
to their starting.

Change-Id: I131f52ce2a3d316d608d209d91260b209a035145