Neutron 8to9 upgrade workflow (SOC-10080)
Prior to SOC 9 a neutron-vpn-agent service provided VPNaaS capabilities
but as of SOC 9 a neutron-l3-agent service is used instead. Therefore
we need to correctly handle the deprecation of this service.

Additionally when the neutron-upgrade.yml playbook is run for the first
time there won't be any neutron-l3-agent services running on NEU-VPNA
nodes so we have to ensure that the status check doesn't fail because
the SOC 9 neutron-l3-agent instance hasn't been deployed yet.

Depends-On: If0f2038593438321bf9bb382fd93d6cd5963bc35
Change-Id: I43672b11c2f30c0f5f80187678b6eba6acafc282