Add policy.d/neutron-fwaas.json.j2 (SOC-10280)
The neutron-fwaas.json file will be rendered from the .j2 file
using the "provided-data" option specified in the neutron-l3-agent
service definition. This json rendering mechanism already exists
in the neutron-ansible playbooks.

Neutron-server errors out if it failed to read any file in the policy.d/
directory. Backup of config files, including policy json files, are
owned by root with permission of 0640. As a result neutron-server will
error out when the neutron-ansible playbook saves a previous version
of a policy json file in policy.d/. This patch also make backup files
retain the attributes of the original copy.

Also made saving old copies of neutron config files a configurable
behavior when _write_conf*.yml is used to render the config files.
Default is still to save the original copy.

Change-Id: I078e5030792315e203d0cac0505a521422f9f230