Add network.target "After" option (bsc#1169770)
This patch fixes a condition with wicked and openvswitch for
servers with a large number of ovs ports. The issue that occurs
is after a reboot, wicked does not correctly setup the interfaces
and the node does not initialize correctly and has no ip addresses.
A followup reboot is necessary to resolve this. Adding the

After=syslog.target network.target

to the neutron-ovs-cleanup unit file resolves the issue. This change is
consistent with a Crowbar SOC cloud, as it also has the option
in the unit file. This change removes the need for a second reboot
to restore networking.

Change-Id: Id07eb95a3e103bea3a3de18ddcb127e2ab205d84
(cherry picked from commit 507fb50c78723a6d34dd16a1267efd19d49a6598)