Allow users to stop monitoring rootwrap daemon (bsc#1182317)
There are times when the neutron agents on a host don't need to use
rootwrap for extended periods of time, e.g. a controller with no dhcp
nets or routers. A rootwrap daemon self terminates if it goes unused for
600s by default, and a neutron agent just spawns a new one if the old
one is no longer running. But monasca raises an alarm if there are no
rootwraps running. This is incorrect as there is no problem and nothing
for the operator to do.

The monasca neutron detection plugin includes neutron-rootwrap in
process_names list by default. This patch allows users to override
process_names by setting it in the monasca_agent_plugin_args variable
in roles/neutron-common/defaults/main.yml.

Change-Id: Ied0f350dbf518b864c9e2d31dcba08ff19a1b600